# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KenyaVet Domestic Staff Vetting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vetting-requests:
    get:
      operationId: listVettingRequests
      tags: [vetting-requests]
      summary: List vetting requests for the current employer
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Vetting request list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VettingRequestList"

    post:
      operationId: createVettingRequest
      tags: [vetting-requests]
      summary: Submit a new vetting request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVettingRequestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VettingRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vetting-requests/{id}:
    get:
      operationId: getVettingRequest
      tags: [vetting-requests]
      summary: Get a vetting request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vetting request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VettingRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateVettingRequest
      tags: [vetting-requests]
      summary: Update a vetting request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVettingRequestBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VettingRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vetting-requests/{id}/steps:
    get:
      operationId: getVettingSteps
      tags: [vetting-requests]
      summary: Get step-by-step progress for a vetting request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vetting steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VettingStep"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── REFERENCES ──────────────────────────────────────────────────────────────
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: Search vetted worker database
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: query
          in: query
          schema:
            type: string
        - name: minScore
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Worker search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerList"
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get a vetted worker profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workers/verify/{qrCode}:
    get:
      operationId: verifyWorkerByQr
      tags: [workers]
      summary: Verify a worker by their QR code
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerVerification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── REPORTS ─────────────────────────────────────────────────────────────────
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List current staff for the employer
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffRecord"

    post:
      operationId: addStaff
      tags: [staff]
      summary: Add a staff member to the employer roster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffRecord"
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffRecord"

    delete:
      operationId: removeStaff
      tags: [staff]
      summary: Remove a staff member from roster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── ADMIN ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── AUTH ──────────────────────────────────────────────────────────────────
...
    CreateVettingRequestBody:
      type: object
      required: [workerName, workerPhone, workerIdNumber, workerRole, packageId]
      properties:
        workerName:
          type: string
        workerPhone:
          type: string
        workerIdNumber:
          type: string
        workerRole:
          type: string
        packageId:
          type: integer
        workerPhotoUrl:
          type: ["string", "null"]
        workerAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    UpdateVettingRequestBody:
      type: object
      properties:
        workerName:
          type: ["string", "null"]
        workerPhone:
          type: ["string", "null"]
        workerAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    VettingRequest:
      type: object
      required: [id, employerId, workerName, workerPhone, workerIdNumber, workerRole, packageId, packageName, status, trustScore, createdAt]
      properties:
        id:
          type: integer
        employerId:
          type: integer
        workerName:
          type: string
        workerPhone:
          type: string
        workerIdNumber:
          type: string
        workerRole:
          type: string
        packageId:
          type: integer
        packageName:
          type: string
        status:
          type: string
          enum: [pending_payment, paid, in_progress, completed, cancelled]
        trustScore:
          type: ["integer", "null"]
        workerPhotoUrl:
          type: ["string", "null"]
        workerAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        reportId:
          type: ["integer", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    VettingRequestList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VettingRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    VettingStep:
      type: object
      required: [id, vettingRequestId, stepName, stepKey, status, order]
      properties:
        id:
          type: integer
        vettingRequestId:
          type: integer
        stepName:
          type: string
        stepKey:
          type: string
          enum: [identity_check, reference_calls, dci_certificate, social_media_review, address_visit, report_generation]
        status:
          type: string
          enum: [pending, in_progress, completed, failed, skipped]
        order:
          type: integer
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]

    # ─── REFERENCES ────────────────────────────────────────────────────────────
...
    Worker:
      type: object
      required: [id, name, role, trustScore, verifiedAt, qrCode, vetCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        trustScore:
          type: integer
        photoUrl:
          type: ["string", "null"]
        neighbourhood:
          type: ["string", "null"]
        yearsExperience:
          type: ["integer", "null"]
        languages:
          type: array
          items:
            type: string
        verifiedAt:
          type: string
        qrCode:
          type: string
        vetCount:
          type: integer
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
    WorkerList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Worker"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    WorkerVerification:
      type: object
      required: [worker, isValid, verifiedAt]
      properties:
        worker:
          $ref: "#/components/schemas/Worker"
        isValid:
          type: boolean
        verifiedAt:
          type: string

    # ─── REPORTS ───────────────────────────────────────────────────────────────
...
    CreateStaffBody:
      type: object
      required: [name, role, startDate]
      properties:
        name:
          type: string
        role:
          type: string
        startDate:
          type: string
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        vettingRequestId:
          type: ["integer", "null"]
    UpdateStaffBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]
    StaffRecord:
      type: object
      required: [id, employerId, name, role, startDate, active, createdAt]
      properties:
        id:
          type: integer
        employerId:
          type: integer
        name:
          type: string
        role:
          type: string
        startDate:
          type: string
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        vettingRequestId:
          type: ["integer", "null"]
        trustScore:
          type: ["integer", "null"]
        renewalDueAt:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── ADMIN ─────────────────────────────────────────────────────────────────
...